+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Project reports includes week number #975
 - Reports performance issues #984
 - Reports infinite loops #994
+- Unconfirmed hours included in summary view #1031
 
 ### Changed
 - Summary report shows current week #980